# pkg/api/ipvs/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "IPVS public API",
    "version": "2.0",
    "contact": {
      "name": "E",
      "url": "https://blog.bullgare.com/2020/07/complete-list-of-swagger-options-to-protobuf-file/"
    }
  },
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v2/ipvs/real-servers/update": {
      "post": {
        "summary": "Update real servers for one IP-virtual server",
        "operationId": "IpvsAdmin_UpdateRealServers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ipvsUpdateRealServersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ipvsUpdateRealServersRequest"
            }
          }
        ],
        "tags": [
          "IpvsAdmin"
        ]
      }
    },
    "/v2/ipvs/virtual-server/find": {
      "post": {
        "summary": "Find IP-virtual server by its identity",
        "operationId": "IpvsAdmin_FindVirtualServer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ipvsFindVirtualServerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ipvsFindVirtualServerRequest"
            }
          }
        ],
        "tags": [
          "IpvsAdmin"
        ]
      }
    },
    "/v2/ipvs/virtual-servers/list": {
      "post": {
        "summary": "List all IP-virtual servers with/without theirs real servers",
        "operationId": "IpvsAdmin_ListVirtualServers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ipvsListVirtualServersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ipvsListVirtualServersRequest"
            }
          }
        ],
        "tags": [
          "IpvsAdmin"
        ]
      }
    },
    "/v2/ipvs/virtual-servers/update": {
      "post": {
        "summary": "Update IP-virtual servers",
        "operationId": "IpvsAdmin_UpdateVirtualServers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ipvsUpdateVirtualServersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ipvsUpdateVirtualServersRequest"
            }
          }
        ],
        "tags": [
          "IpvsAdmin"
        ]
      }
    }
  },
  "definitions": {
    "IssueReasonCode": {
      "type": "string",
      "enum": [
        "ExternalError",
        "Unsupported",
        "VirtualServerNotFound",
        "RealServerNotFound"
      ],
      "default": "ExternalError",
      "title": "- ExternalError: external error that happens out from external libs\n - Unsupported: Something is not supported by IPVS implementor\n - VirtualServerNotFound: when delete/update VS is not exist subject\n - RealServerNotFound: when delete/update RS is not exist subject"
    },
    "ipvsFindVirtualServerRequest": {
      "type": "object",
      "properties": {
        "virtualServerIdentity": {
          "$ref": "#/definitions/ipvsVirtualServerIdentity"
        },
        "includeReals": {
          "type": "boolean",
          "title": "includeReals add real servers into response"
        }
      },
      "title": "FindVirtualServerRequest it gets-or-fails the IP-virtual server by its identity"
    },
    "ipvsFindVirtualServerResponse": {
      "type": "object",
      "properties": {
        "virtualServer": {
          "$ref": "#/definitions/ipvsVirtualServerWithReals"
        }
      },
      "title": "FindVirtualServerResponse response with virtual server with/without their real server(s)"
    },
    "ipvsIssueReason": {
      "type": "object",
      "properties": {
        "code": {
          "$ref": "#/definitions/IssueReasonCode",
          "title": "issue reason code"
        },
        "message": {
          "type": "string",
          "title": "issue reason in textual format"
        }
      },
      "title": "IssueReason issue reason on update/delete operation"
    },
    "ipvsListVirtualServersRequest": {
      "type": "object",
      "properties": {
        "includeReals": {
          "type": "boolean",
          "title": "includeReals add real servers into response"
        }
      },
      "title": "ListVirtualServersRequest ask to list all virtual servers with/without its reals"
    },
    "ipvsListVirtualServersResponse": {
      "type": "object",
      "properties": {
        "virtualServers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ipvsVirtualServerWithReals"
          }
        }
      },
      "title": "ListVirtualServersResponse list all virtual servers with/without its reals"
    },
    "ipvsNetworkTransport": {
      "type": "string",
      "enum": [
        "TCP",
        "UDP"
      ],
      "default": "TCP",
      "title": "NetworkTransport is an IP-network type transport to use with instance of VirtualServer"
    },
    "ipvsPacketFwdMethod": {
      "type": "string",
      "enum": [
        "DirectRouting",
        "Tunnel",
        "Masquerading"
      ],
      "default": "DirectRouting",
      "description": "- DirectRouting: DirectRouting is direct routing\n - Tunnel: Tunnel uses ipip encapsulation\n - Masquerading: Masquerading is network access translation (NAT)",
      "title": "PacketFwdMethod represents method of forwarding packets in VS from user to real server(s)"
    },
    "ipvsRealServer": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/ipvsRealServerAddress"
        },
        "packetForwarder": {
          "$ref": "#/definitions/ipvsPacketFwdMethod",
          "title": "packet_forwarder use one of (DR | TUN | NAT) packet forward method"
        },
        "weight": {
          "type": "integer",
          "format": "int64",
          "title": "Weight is an integer specifying the capacity of a server relative to the others in the pool.\nThe valid values of weight are 0 through to 65535. The default is 1.\nQuiescent servers are specified with a weight of zero. A quiescent server will receive no new jobs\nbut still serve the existing jobs, for all scheduling algorithms distributed with the Linux Virtual Server.\nSetting a quiescent server may be useful if the server is overloaded or needs to be\ntaken out of service for maintenance"
        },
        "upperThreshold": {
          "type": "integer",
          "format": "int64",
          "title": "upper_threshold is an integer specifying the upper connection threshold of a server.\nThe valid values of uthreshold are 0 through to 65535. The default is 0,\nwhich means the upper connection threshold is not set. If uthreshold is set with other values,\nno new connections will be sent to the server when the number of its connections exceeds its\nupper connection threshold"
        },
        "lowerThreshold": {
          "type": "integer",
          "format": "int64",
          "title": "lower_threshold is an integer specifying the lower connection threshold of a server.\nThe valid values of lthreshold are 0 through to 65535. The default is 0, which means the lower\nconnection threshold is not set. If lthreshold is set with other values, the server will receive\nnew connections when the number of its connections drops below its lower connection threshold.\nIf lthreshold is not set but uthreshold is set, the server will receive new connections when the\nnumber of its connections drops below three forth of its upper connection threshold"
        }
      },
      "title": "RealServer is the real server"
    },
    "ipvsRealServerAddress": {
      "type": "object",
      "properties": {
        "host": {
          "type": "string",
          "title": "host is IP or host name"
        },
        "port": {
          "type": "integer",
          "format": "int64",
          "title": "port is port \u003e 0"
        }
      },
      "title": "RealServerAddress the IP-network real server address"
    },
    "ipvsRealServerIssue": {
      "type": "object",
      "properties": {
        "reason": {
          "$ref": "#/definitions/ipvsIssueReason"
        },
        "delete": {
          "$ref": "#/definitions/ipvsRealServerAddress",
          "title": "when-delete on delete operation"
        },
        "update": {
          "$ref": "#/definitions/ipvsRealServer",
          "title": "when-update on update operation"
        }
      },
      "title": "RealServerIssue issue that happens on update/delete real server"
    },
...
    "ipvsUpdateRealServersRequest": {
      "type": "object",
      "properties": {
        "forceUpsert": {
          "type": "boolean"
        },
        "virtualServerIdentity": {
          "$ref": "#/definitions/ipvsVirtualServerIdentity"
        },
        "delete": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ipvsRealServerAddress"
          }
        },
        "update": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ipvsRealServer"
          }
        }
      },
      "title": "UpdateRealServersRequest request foe delete+update real server(s)"
    },
    "ipvsUpdateRealServersResponse": {
      "type": "object",
      "properties": {
        "issues": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ipvsRealServerIssue"
          }
        }
      },
      "title": "UpdateRealServersResponse collects issues on update real servers"
    },
    "ipvsUpdateVirtualServersRequest": {
      "type": "object",
      "properties": {
        "forceUpsert": {
          "type": "boolean"
        },
        "delete": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ipvsVirtualServerIdentity"
          }
        },
        "update": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ipvsVirtualServer"
          }
        }
      },
      "title": "UpdateVirtualServersRequest request for delete+update virtual server(s)"
    },
    "ipvsUpdateVirtualServersResponse": {
      "type": "object",
      "properties": {
        "issues": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ipvsVirtualServerIssue"
          }
        }
      },
      "title": "UpdateRealServersResponse collects issues on update IP-virtual servers"
    },
    "ipvsVirtualServer": {
      "type": "object",
      "properties": {
        "identity": {
          "$ref": "#/definitions/ipvsVirtualServerIdentity"
        },
        "scheduleMethod": {
          "$ref": "#/definitions/ipvsScheduleMethod"
        }
      },
      "title": "VirtualServer an IP virtual server are is part od IPVS arch"
    },
    "ipvsVirtualServerAddress": {
      "type": "object",
      "properties": {
        "network": {
          "$ref": "#/definitions/ipvsNetworkTransport"
        },
        "host": {
          "type": "string"
        },
        "port": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "VirtualServerAddress represents IP network address of virtual server"
    },
    "ipvsVirtualServerIdentity": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/ipvsVirtualServerAddress"
        },
        "firewallMark": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "VirtualServerIdentity virtual server may be identified by address or firewall-mark"
    },
    "ipvsVirtualServerIssue": {
      "type": "object",
      "properties": {
        "reason": {
          "$ref": "#/definitions/ipvsIssueReason"
        },
        "delete": {
          "$ref": "#/definitions/ipvsVirtualServerIdentity",
          "title": "when-delete on delete operation"
        },
        "update": {
          "$ref": "#/definitions/ipvsVirtualServer",
          "title": "when-update on update operation"
        }
      },
      "title": "VirtualServerIssue issue that happens on update/delete IP-virtual server"
    },
    "ipvsVirtualServerWithReals": {
      "type": "object",
      "properties": {
        "virtualServer": {
          "$ref": "#/definitions/ipvsVirtualServer"
        },
        "realServers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ipvsRealServer"
          }
        }
      },
      "title": "VirtualServerWithReals IP-virtual server and associated its real IP servers"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}